static model import
