--- src/models/User.js
@@ -1,35 +1,0 @@
-import { DataTypes } from "sequelize";
-import sequelize from "../lib/sequelize.js";
-
-const User = sequelize.define("User", {
-  id: {
-    type: DataTypes.UUID,
-    primaryKey: true,
-    allowNull: false,
-    defaultValue: DataTypes.UUIDV4,
-  },
-  name: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-  email: {
-    type: DataTypes.STRING,
-    allowNull: false,
-    unique: true,
-  },
-  created_at: {
-    type: DataTypes.DATE,
-    allowNull: false,
-    defaultValue: DataTypes.NOW
-  },
-  updated_at: {
-    type: DataTypes.DATE,
-    allowNull: false,
-    defaultValue: DataTypes.NOW,
-  }
-}, {
-  tableName: "User",
-  underscored: true,
-});
-
-export default User;
