+++ src/models/index.js
@@ -1,32 +1,12 @@
-import fs from "fs";
-import path from "path";
-import { DataTypes } from "sequelize";
-import { fileURLToPath } from "url";
-import sequelize from "../lib/sequelize.js";
+import sequelize from "../lib/sequelize.js"
+import UserModel from "../modules/users/user.model.js";
 
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-const basename = path.basename(__filename);
+const User = UserModel(sequelize, sequelize.Sequelize.DataTypes);
 
-const db = {};
-
-const files = fs.readdirSync(__dirname).filter(
-  (file) => file.indexOf(".") !== 0 && file !== basename && file.slice(-3) === ".js"
-);
-
-for(let file of files){
-  const modelImport = await import (path.join(__dirname, file));
-  const model = modelImport.default(sequelize, DataTypes);
-  db[model.name] = model;
-}
-
-Object.keys(db).forEach((modelName)=>{
-  if(db[modelName].associate){
-    db[modelName].associate(db);
-  }
-});
-
-db.sequelize = sequelize;
-db.Sequelize = sequelize.Sequelize;
+const db = {
+  sequelize,
+  Sequelize: sequelize.Sequelize,
+  User,
+};
 
 export default db;
