+++ src/modules/users/user.model.js
@@ -1,0 +1,26 @@
+export default (sequelize, DataTypes) => {
+  const User = sequelize.define("User", {
+    id: {
+      type: DataTypes.UUID,
+      primaryKey: true,
+      allowNull: false,
+      defaultValue: DataTypes.UUIDV4,
+    },
+    name: {
+      type: DataTypes.STRING,
+      allowNull: false,
+    },
+    email: {
+      type: DataTypes.STRING,
+      allowNull: false,
+      unique: true,
+    },
+  }, {
+    tableName: "User",
+    underscored: true,
+    timestamps: true,
+  });
+
+  return User;
+}
+
